devai-api/requirements.txt
fastapi==0.109.2
google-cloud-aiplatform==1.48.0
google-auth==2.27.0
itsdangerous==2.1.2
jinja2==3.1.4
langchain==0.2.10
langchain_google_vertexai==1.0.5
markdown==3.5.2
types-Markdown==3.5.0.20240129
uvicorn[standard]==0.27.0.post1
python-multipart==0.0.18
langsmith==0.1.129
google-generativeai==0.4.0
python-gitlab==4.4.0

jira==3.8.0
pydantic==2.6.1
atlassian-python-api==3.41.10
langchain_community==0.2.10

click==8.1.7
PyGithub==2.5.0
GitPython==3.1.43